Purchase item

+++ server/Infrastructure/Migrations/20230615042107_InitialCreate.cs
@@ -29,7 +29,7 @@ namespace Infrastructure.Migrations
                     Gender = table.Column<bool>(type: "tinyint(1)", nullable: false),
                     Dob = table.Column<DateTime>(type: "datetime(6)", nullable: false),
                     Address = table.Column<string>(type: "longtext", nullable: false),
-                    Points = table.Column<int>(type: "int", nullable: false),
+                    Points = table.Column<double>(type: "double", nullable: false),
                     Role = table.Column<int>(type: "int", nullable: false)
                 },
                 constraints: table =>
